Second version of code.
Main scene has been added to the already functional intro scene.
With the help of Vuforia core samples, user defined target is used augment the Car model inside the environment.
In this version of code, we can manipulate with only 1 of the 4 car models (spyder Lamborghini).
We could open or close Driver's door, start or stop the engine and change the color of the car.